macros.cfg - START_PRINT: Changed start of print
Stuff that happens at start of print:
Home all.
Move to X3 Y200. The nozzle is right beside the glass bed in this position.
Set extruder and bed temperature for the first layer and wait. (After the first layer the temp is whatever was set in the slicer)
Move to Y20 and extrude 10mm.
Move to X10 where the nozzle is above the glass by a paper thin amount wiping away excess filament.